ABIOS-2931 Changed UNIX timestamp type from Int to Double

+++ Ambatana/src/Helper/FoundationExtensions/Date+LG.swift
@@ -180,7 +180,7 @@ extension Date {
         return Calendar.current.component(.year, from: self)
     }
     
-    var millisecondsSince1970: Int {
-        return Int((self.timeIntervalSince1970 * 1000.0).rounded())
+    var millisecondsSince1970: Double {
+        return (self.timeIntervalSince1970 * 1000.0).rounded()
     }
 }
